Remove old quickstart
