+++ iam/api/QuickStart/QuickStart.csproj
@@ -1,9 +1,10 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp2.1</TargetFramework>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Google.Apis.Iam.v1" Version="1.45.0.1904" />
+    <PackageReference Include="Google.Apis.Cloudresourcemanager.v1" Version="1.43.0.1859" />
+    <PackageReference Include="Google.Apis.Iam.v1" Version="1.43.0.1835" />
   </ItemGroup>
 </Project>
