--- iam/api/QuickStartV2/README.md
@@ -1,40 +1,0 @@
-# IAM Quickstart Sample
-
-This sample demonstrates how to get started using the Google Cloud IAM APIs
-in C#.
-
-This sample requires [.NET Core 2.1](https://www.microsoft.com/net/core) or
-later.  That means using [Visual Studio 2017](https://www.visualstudio.com/), 
-or the command line. 
-
-## Links
-
-- [Cloud IAM Docs](https://cloud.google.com/iam/docs/)
-
-## Build and Run
-
-1.  **Follow the set-up instructions in [the documentation](https://cloud.google.com/dotnet/docs/setup).**
-
-1.  Enable APIs for your project.
-    [Click here](https://console.cloud.google.com/flows/enableapi?apiid=iam.googleapis.com&showconfirmation=true)
-    to visit Cloud Platform Console and enable the Google Cloud IAM API.
-
-1.  Replace the placeholder information in QuickStart.cs with your project-specific information. 
-
-1.  From the command line:
-
-    ```bash
-    dotnet restore
-    dotnet run
-    ```
-
-    The application should grant the member you inputted the Logs viewer role, print the permissions that
-    they have as a Logs viewer, then revoke their Logs viewer role.
-
-## Contributing changes
-
-* See [CONTRIBUTING.md](../../../CONTRIBUTING.md)
-
-## Licensing
-
-* See [LICENSE](../../../LICENSE)
